requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Jsonpify==1.5.0
flask-marshmallow==0.15.0
Flask-Migrate==2.6.0
Flask-RESTful==0.3.10
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
referencing==0.30.0
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
